+++ src/data/db/migrationRunner.ts
@@ -18,6 +18,7 @@ import * as migration015 from './migrations/015_lore_entity';
 import * as migration016 from './migrations/016_spoiler_safe';
 import * as migration017 from './migrations/017_dnf_reflection';
 import * as migration018 from './migrations/018_shelf_book_index';
+import * as migration019 from './migrations/019_reading_run';
 
 const log = createLogger('db/migrations');
 
@@ -56,6 +57,7 @@ const migrations: Migration[] = [
   { version: migration016.version, up: migration016.up },
   { version: migration017.version, up: migration017.up },
   { version: migration018.version, up: migration018.up },
+  { version: migration019.version, up: migration019.up },
 ];
 
 export const LATEST_SCHEMA_VERSION = migrations[migrations.length - 1]?.version ?? 0;